Update color naming
I've changed the naming conventions to broader terms. I'll be working on documentation that will include definition of where and how to use the color (similar to what is already listed above). This naming structure will givesus more flexibility in the future. 

Purely for example, if we change our neutral set of grays to a neutral set of blues there are no issues with naming, let alone possible confusion with what is "primary" or " secondary." This type of definition is more useful in the documentation than the actual name.